requirements.txt
amqp==2.4.2
asn1crypto==0.24.0
backoff==1.8.0
bcrypt==3.1.6
boto3==1.9.154
botocore==1.12.154
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.4.2
docutils==0.14
elasticsearch==7.0.1
Fabric3==1.14.post1
future==0.17.1
idna==2.8
jmespath==0.9.4
kombu==4.5.0
paramiko==2.4.2
prompt-toolkit==1.0.16
pyasn1==0.4.5
pycparser==2.19
Pygments==2.4.0
PyNaCl==1.3.0
python-dateutil==2.8.0
PyYAML==5.1
redis==3.2.1
requests==2.22.0
s3transfer==0.2.0
six==1.12.0
tqdm==4.32.1
urllib3==1.24.3
vine==1.3.0
wcwidth==0.1.7
